requirements.txt
geopy
osmium
pyyaml
shapely
mysql-connector-python
